fix: PN5180 SPI transfer loops used a uint8_t counter against 508-byte frames
transceiveCommand's four transfer loops counted with uint8_t while both
buffer lengths are size_t and legal frames run to 508 bytes (readData
accepts up to 508; RX_BYTES_RECEIVED_MASK allows 511). Any transfer of
256+ bytes wrapped the counter at 255 and spun forever, hanging the scan
task inside the SPI transaction until the 30s task watchdog panicked the
board. A corrupted RX_STATUS length on marginal wiring is enough to ask
for such a read, which matches spontaneous task_wdt reboots observed on
a shared-bus C6 bench rig.

+++ lib/PN5180/PN5180.cpp
@@ -611,7 +611,7 @@ bool PN5180::transceiveCommand(uint8_t *sendBuffer, size_t sendBufferLen, uint8_
   }
 #ifdef DEBUG
   PN5180DEBUG(F("Sending SPI frame: '"));
-  for (uint8_t i=0; i<sendBufferLen; i++) {
+  for (size_t i=0; i<sendBufferLen; i++) {
     if (i>0) PN5180DEBUG(" ");
     PN5180DEBUG(formatHex(sendBuffer[i]));
   }
@@ -632,7 +632,7 @@ bool PN5180::transceiveCommand(uint8_t *sendBuffer, size_t sendBufferLen, uint8_
   // 1.
   digitalWrite(PN5180_NSS, LOW); delay(2);
   // 2.
-  for (uint8_t i=0; i<sendBufferLen; i++) {
+  for (size_t i=0; i<sendBufferLen; i++) {
     pn5180_spi.transfer(sendBuffer[i]);
   }
   // 3.
@@ -669,7 +669,7 @@ bool PN5180::transceiveCommand(uint8_t *sendBuffer, size_t sendBufferLen, uint8_
   // 1.
   digitalWrite(PN5180_NSS, LOW); delay(2);
   // 2.
-  for (uint8_t i=0; i<recvBufferLen; i++) {
+  for (size_t i=0; i<recvBufferLen; i++) {
     recvBuffer[i] = pn5180_spi.transfer(0xff);
   }
   // 3.
@@ -700,7 +700,7 @@ bool PN5180::transceiveCommand(uint8_t *sendBuffer, size_t sendBufferLen, uint8_
 
 #ifdef DEBUG
   PN5180DEBUG(F("Received: "));
-  for (uint8_t i=0; i<recvBufferLen; i++) {
+  for (size_t i=0; i<recvBufferLen; i++) {
     if (i > 0) PN5180DEBUG(" ");
     PN5180DEBUG(formatHex(recvBuffer[i]));
   }
